qr and emp circle api

+++ app/Traits/WriteLogger.php
@@ -6,12 +6,14 @@ use Log;
 
 trait WriteLogger
 {
-    function writeLog($channel, $key, $data)
+    function writeLog($channel, $key, $data, $delete = true)
     {
         if (config("app.WRITE_LOG")) {
-            $logFilePath = storage_path("logs/{$channel}.log");
-            if (file_exists($logFilePath)) {
-                unlink($logFilePath);
+            if ($delete) {
+                $logFilePath = storage_path("logs/{$channel}.log");
+                if (file_exists($logFilePath)) {
+                    unlink($logFilePath);
+                }
             }
             $data = ['key' => $key, "data" => $data];
             Log::channel($channel)->info($data);
